chore: update parameter name
